requirements.txt
flask
python-dotenv
gunicorn
psycopg2
pandas
openpyxl